add auth checks to server actions (is fine without it due to rls but double security)

+++ src/server/actions/deletePlan.ts
@@ -4,6 +4,16 @@ import { redirect } from "next/navigation"
 
 export default async function deletePlan(planId: string) {
   const supabase = createSupabaseServerClient()
+
+  // check auth
+  const {
+    data: { user },
+  } = await supabase.auth.getUser()
+
+  if (!user) {
+    throw new Error("User not found")
+  }
+
   await supabase.from("plan").delete().eq("id", planId)
   redirect("/plans")
 }
